# remove unnecessary bottom navigation from workspace tab